requirements.txt
flask
pandas
pyodbc
schedule
fastapi
uvicorn
matplotlib
python-pptx